# api.yml
openapi: 3.0.3
info:
  title: Clinic Patient Management API
  description: API for managing users, authentication, doctors, medication, appointments, allergies, notifications, and reports.
  version: 1.0.0
servers:
  - url: https://ehrbackend-production-a3a0.up.railway.app/api
paths:
...
  /auth/register:
    post:
      summary: Register a new user (patient or doctor)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthResponse'
                  - $ref: '#/components/schemas/UserResponseDTO'
        '500':
          description: Internal error during registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '500':
          description: Internal error during login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify:
    get:
      summary: Verify user email using token
      tags: [Auth]
      parameters:
        - in: query
          name: token
          schema:
            type: string
          required: true
          description: Verification token sent via email
      responses:
        '200':
          description: Email verified successfully
          content:
            text/plain:
              schema:
                type: string
                example: Email verified successfully. You can login now.
        '400':
          description: Invalid or expired verification token
          content:
            text/plain:
              schema:
                type: string
                example: Invalid verification token
  /auth/refresh-token:
    post:
      summary: Refresh JWT access token using refresh token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
      responses:
        '200':
          description: Fresh token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            text/plain:
              schema:
                type: string
                example: Invalid or Expired token

  # --------------------------------------------------------------------
  # Users
  # --------------------------------------------------------------------
...
  /patients/{patientId}/diet-instructions:
    get:
      summary: Get diet instructions for a patient
      tags: [Medication]
      parameters:
        - in: path
          name: patientId
          schema:
            type: string
          required: true
          description: Patient ID (ehrId)
      responses:
        '200':
          description: List of diet instructions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DietInstruction'
        '404':
          description: Patient not found
...
  /patient/{patient_id}/diet_instruction:
    post:
      summary: Add a diet instruction for a patient
      tags: [Diet Instructions]
      parameters:
        - in: path
          name: patient_id
          schema:
            type: string
          required: true
          description: Patient ID (ehrId)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DietInstructionRequest'
      responses:
        '200':
          description: Diet instruction created
        '400':
          description: Invalid input

    get:
      summary: Get a diet instruction by ID
      tags: [Diet Instructions]
      parameters:
        - in: path
          name: patient_id
          schema:
            type: string
          required: true
          description: Patient ID (ehrId)
        - in: query
          name: dietId
          schema:
            type: integer
            format: int64
          required: true
          description: Diet instruction ID
      responses:
        '200':
          description: Diet instruction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DietInstruction'
        '404':
          description: Diet instruction not found

    delete:
      summary: Delete a diet instruction
      tags: [Diet Instructions]
      parameters:
        - in: path
          name: patient_id
          schema:
            type: string
          required: true
          description: Patient ID (ehrId)
        - in: query
          name: dietId
          schema:
            type: integer
            format: int64
          required: true
          description: Diet instruction ID
      responses:
        '200':
          description: Diet instruction deleted
        '404':
          description: Diet instruction not found
  /patient/{patient_id}/{dietId}/diet_instruction:
    put:
      summary: Update a diet instruction
      tags: [Diet Instructions]
      parameters:
        - in: path
          name: patient_id
          schema:
            type: string
          required: true
          description: Patient ID (ehrId)
        - in: path
          name: dietId
          schema:
            type: integer
            format: int64
          required: true
          description: Diet instruction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DietInstructionRequest'
      responses:
        '200':
          description: Diet instruction updated
        '404':
          description: Diet instruction not found

  # --------------------------------------------------------------------
  # Notifications
  # --------------------------------------------------------------------
...
  /restrictions/request:
    post:
      summary: Request restriction grant
      tags: [Restrictions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestrictionRequestDto'
      responses:
        '200':
          description: Restriction request created
  /restrictions/approve/{requestId}:
    put:
      summary: Approve restriction grant
      tags: [Restrictions]
      parameters:
        - in: path
          name: requestId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: Restriction approved
  /restrictions/reject/{requestId}:
    put:
      summary: Reject restriction grant
      tags: [Restrictions]
      parameters:
        - in: path
          name: requestId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Restriction rejected
  /restrictions/patient/{patientId}:
    get:
      summary: Get restrictions for a patient
      tags: [Restrictions]
      parameters:
        - in: path
          name: patientId
          schema:
            type: string
          required: true
          description: Patient ID (ehrId)
      responses:
        '200':
          description: List of restrictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestrictionGrantResponseDto'
components:
  schemas:
...
    UserRequestDTO:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        dob:
          type: string
          description: Date in mm-dd-yyyy format
          example: 07-09-1990
        password:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        phoneNo:
          type: integer
          format: int64
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'
        gender:
          $ref: '#/components/schemas/Gender'
        emergencyContact:
          $ref: '#/components/schemas/Emergency'
        doctorProfile:
          $ref: '#/components/schemas/DoctorProfile'
        maritalStatus:
          type: string
    UserResponseDTO:
      type: object
      properties:
        email:
          type: string
        ehrId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNo:
          type: integer
          format: int64
        dob:
          type: string
          description: Date in mm-dd-yyyy format
          example: 07-09-1990
        gender:
          $ref: '#/components/schemas/Gender'
        role:
          $ref: '#/components/schemas/Role'
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'
        address:
          $ref: '#/components/schemas/Address'
        doctorProfile:
          $ref: '#/components/schemas/DoctorProfile'
        maritalStatus:
          type: string
        emergencyContact:
          $ref: '#/components/schemas/Emergency'
    Address:
      type: object
      properties:
        addressLine:
          type: string
        state:
          type: string
        city:
          type: string
        pin_code:
          type: integer
          format: int64
          description: Postal code (length 6)
    Emergency:
      type: object
      properties:
        contactName:
          type: string
        contactPhoneNo:
          type: integer
          format: int64
          description: Contact phone number (length 10)
        relationship:
          type: string
    User:
      type: object
      properties:
        ehrId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
          description: True if the user's email has been verified
        dob:
          type: string
          description: Date in mm-dd-yyyy format
        maritalStatus:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        phoneNo:
          type: integer
          format: int64
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          $ref: '#/components/schemas/Address'
        emergencyContact:
          $ref: '#/components/schemas/Emergency'
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'

    # -----------------------------------------------------------------
    # Auth
    # -----------------------------------------------------------------
    AuthResponse:
      type: object
      properties:
        userResponseDTO:
          $ref: '#/components/schemas/UserResponseDTO'
        tokenResponse:
          $ref: '#/components/schemas/TokenResponse'
    AuthLoginResponse:
      type: object
      properties:
        loginResponse:
          $ref: '#/components/schemas/LoginResponse'
        tokenResponse:
          $ref: '#/components/schemas/TokenResponse'
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        status:
          type: boolean
        message:
          type: string
        userResponseDTO:
          $ref: '#/components/schemas/UserResponseDTO'
    TokenResponse:
      type: object
      properties:
        generateAccessToken:
          type: string
        generateRefreshToken:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer

    # -----------------------------------------------------------------
    # Doctors
    # -----------------------------------------------------------------
    Doctor:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/User'
        doctorProfile:
          $ref: '#/components/schemas/DoctorProfile'
    DoctorProfile:
      type: object
      properties:
        aadhaarNumber:
          type: string
        panNumber:
          type: string
        specialization:
          type: string
        licenseNumber:
          type: string
        aboutDoctor:
          type: string
        degrees:
          type: array
          items:
            type: string
...
    DietInstruction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name_of_diet:
          type: string
        food_description:
          type: string
        prescribeBy:
          $ref: '#/components/schemas/Doctor'
        patientId:
          $ref: '#/components/schemas/User'
    DietInstructionRequest:
      type: object
      properties:
        name_of_diet:
          type: string
        food_description:
          type: string
        prescribeBy:
          type: string
          description: Doctor ID (ehrId)

    # -----------------------------------------------------------------
    # Notifications
    # -----------------------------------------------------------------
...
    RestrictionRequestDto:
      type: object
      properties:
        patientId:
          type: string
        accessorId:
          type: string
        restrictedAttributes:
          type: array
          items:
            type: string
    RestrictionGrantResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientId:
          type: string
        patientName:
          type: string
        accessorId:
          type: string
        accessorName:
          type: string
        status:
          $ref: '#/components/schemas/RestrictionGrantStatus'
        restrictedAttributes:
          type: array
          items:
            type: string
        requestedAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time

    # -----------------------------------------------------------------
    # Enums
    # -----------------------------------------------------------------
    Role:
      type: string
      enum: [PATIENT, DOCTOR, ADMIN]
    Gender:
      type: string
      enum: [MALE, FEMALE, Non_binary, Prefer_not_to_say]
    BloodGroup:
      type: string
      enum: [A_POSITIVE, A_NEGATIVE, B_POSITIVE, B_NEGATIVE, AB_POSITIVE, AB_NEGATIVE, O_POSITIVE, O_NEGATIVE]
...
    RestrictionGrantStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]